build 20
config now saves to the server

you can view the changed data live in the server with the /sp-show
command

added commands to agree with all the settings in the config.

This means you can now choose whether to ignore op powers or not and
whether or not you want debug messages displayed.